requirements.txt
# Dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.3.0
openai>=1.6.1
qdrant-client==1.7.0
langchain-openai>=0.0.5
sentence-transformers==2.2.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3

# Procesamiento de documentos
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0
openpyxl==3.1.2
python-pptx==0.6.21

# Soporte mejorado para archivos .doc antiguos
docx2txt==0.8
# python-magic-bin==0.4.14  # Comentado por problemas de instalación
chardet==5.2.0

# Audio processing
pydub==0.25.1
SpeechRecognition==3.10.0

# Utilidades
tqdm==4.66.1
pathlib2==2.3.7
streamlit==1.29.0

# Dependencias opcionales (descomentadas según necesidad)
# pypandoc==1.11  # Requiere pandoc instalado en el sistema
# antiword==1.4.0  # Para archivos .doc en Linux/Mac
